maint.mk: cache VC_LIST output.
Several syntax checks work on the list of all versioned files.  Regenerating
this list for every such check can be quite time-consuming.  To avoid this,
a wrapper target could generate this list once, perform all checks, and
clean up the list afterwards.

As an added benefit, this allows us to pass -k (--keep-going) to the
sub-make, so that it will perform all syntax checks instead of aborting at
the first one to fail.