maint.mk: Imporved LINGUAS generation.
This change avoids duplicates for en@boldquot and en@quot.  It also ensures
that the list is completely sorted and that any comment lines preceding the
list of linguas will be preserved.  Non-lingua lines following the list of
linguas will be moved in front of the linguas, so it is suggested to avoid
any such lines.  A header should be enough.